+++ filters/ChannelFilter.h
@@ -33,7 +33,7 @@ public:
 	bool getAllChannels() override {return true;}
 	bool getSelectChannels() override {return true;}
 	std::vector<std::wstring> initialize(float sampleRate, unsigned maxFrameCount, std::vector<std::wstring> channelNames) override;
-	void process(float** output, float** input, unsigned frameCount) override;
+	void process(double** output, double** input, unsigned frameCount) override;
 
 private:
 	std::vector<std::wstring> words;
