+++ filters/CopyFilter.h
@@ -47,7 +47,7 @@ public:
 	bool getAllChannels() override {return true;}
 	bool getInPlace() override {return false;}
 	std::vector<std::wstring> initialize(float sampleRate, unsigned maxFrameCount, std::vector<std::wstring> channelNames) override;
-	void process(float** output, float** input, unsigned frameCount) override;
+	void process(double** output, double** input, unsigned frameCount) override;
 
 	std::vector<Assignment> getAssignments() const;
 
@@ -62,7 +62,7 @@ private:
 
 		struct InternalSummand
 		{
-			float factor;
+			double factor;
 			int channel;
 		};
