+++ filters/IIRFilter.h
@@ -29,7 +29,7 @@ public:
 	virtual ~IIRFilter();
 	bool getInPlace() override {return true;}
 	std::vector<std::wstring> initialize(float sampleRate, unsigned maxFrameCount, std::vector<std::wstring> channelNames) override;
-	void process(float** output, float** input, unsigned frameCount) override;
+	void process(double** output, double** input, unsigned frameCount) override;
 
 private:
 	unsigned order;
